+++ src/main/java/edu/it10/dangquangwatch/spring/Repository/AnhtrangsucRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;  
 
 @Repository  
-public interface AnhtrangsucRepository extends CrudRepository<Anhtrangsuc, Long> {}
+public interface AnhtrangsucRepository extends CrudRepository<Anhtrangsuc, Integer> {}
