+++ src/main/java/edu/it10/dangquangwatch/spring/Repository/DonghoRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;  
 
 @Repository  
-public interface DonghoRepository extends CrudRepository<Dongho, Long> {}
+public interface DonghoRepository extends CrudRepository<Dongho, Integer> {}
