+++ src/main/java/edu/it10/dangquangwatch/spring/Repository/TrangsucRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;  
 
 @Repository  
-public interface TrangsucRepository extends CrudRepository<Trangsuc, Long> {}
+public interface TrangsucRepository extends CrudRepository<Trangsuc, Integer> {}
